Set elo column to float type in db

+++ bot/data/db.sql
@@ -76,7 +76,7 @@ CREATE TABLE tcg_leaderboard(
 	discord_id BIGINT NOT NULL,
 	win_count INT DEFAULT 0 NOT NULL,
 	loss_count INT DEFAULT 0 NOT NULL,
-	elo INT DEFAULT 1200 NOT NULL,
+	elo FLOAT DEFAULT 1200 NOT NULL,
 	PRIMARY KEY(discord_id),
 	UNIQUE(discord_id)
 );
